Floating contacts: - font settings moved to Options / Customize / Fonts & Colors - code cleaning; - version bump
git-svn-id: http://svn.miranda-ng.org/main/trunk@3916 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/FloatingContacts/src/filedrop.cpp
@@ -282,8 +282,8 @@ static void SaveFiles( char *szItem, char **ppFiles, int *pnCount )
 			}
 			else
 			{
-				int	 nSize = sizeof(char) * ( strlen( szItem ) + strlen( fd.cFileName ) + sizeof( char ));
-				char *szFile = (char*) malloc( nSize ) ;
+				size_t nSize = sizeof(char) * ( strlen( szItem ) + strlen( fd.cFileName ) + sizeof( char ));
+				char  *szFile = (char*) malloc( nSize ) ;
 				
 				strncpy( szFile, szItem, nSize - 1 );
